backend/requirements.txt
anyio==4.12.0
asn1crypto==1.5.1
attrs==25.4.0
blockfrost-python==0.5.3
cbor2==5.7.1
certifi==2025.11.12
certvalidator==0.11.1
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cose==0.9.dev8
cryptography==46.0.3
ecdsa==0.19.1
ECPy==1.2.5
fastapi==0.109.0
frozendict==2.4.7
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
idna==3.11
mnemonic==0.20
oscrypto==1.3.0
pprintpp==0.4.0
pycardano==0.9.0
pycparser==2.23
pydantic==1.10.17
PyNaCl==1.6.1
python-dotenv==1.0.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.35.1
typeguard==2.13.3
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.27.0
websocket-client==1.9.0
websockets==12.0
torch>=2.0.0
torch-geometric>=2.3.0
networkx>=3.0
